+++ integration/aiohttp_yield_server.py
@@ -2,7 +2,7 @@ import sys; sys.path.insert(0, '.')
 from aiohttp import web
 from asyncio import coroutine
 
-from covador.aiohttp import query_string, json_body, form, params, rparams
+from covador.aiohttp import query_string, json_body, form, params, args
 
 
 @query_string(boo=str)
@@ -23,9 +23,9 @@ def params_view(request, p1, p2):
     return web.Response(text='{0}.{1}'.format(p1, p2))
 
 
-@rparams(boo=str)
+@args(boo=str)
 @coroutine
-def rparams_view(request, boo):
+def args_view(request, boo):
     return web.Response(text=boo)
 
 
@@ -40,7 +40,7 @@ def main():
     app.router.add_get('/qs/', qs_view)
     app.router.add_post('/form/', form_view)
     app.router.add_post('/params/', params_view)
-    app.router.add_get('/rparams/{boo}/', rparams_view)
+    app.router.add_get('/args/{boo}/', args_view)
     app.router.add_post('/json/', json_view)
 
     try: