+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+dev
+===
+
+* [Break] `rparams` renamed to `args`.
+* [Fix] aiohttp and sanic.
+* [Refactor] Real integration tests with whole app starting.
+
+
 0.9.18
 ======